+++ src/Cwaff-Guns/English.cs
@@ -41,7 +41,7 @@ public class English : CwaffGun
           .AddToShop(ItemBuilder.ShopType.Cursula)
           .InitProjectile(GunData.New(clipSize: -1, cooldown: 0.25f, angleVariance: 5.0f, chargeTime: 0f, shootStyle: ShootStyle.Charged, pierceBreakables: true,
             damage: 2.5f, speed: 81.0f, range: 9999f, sprite: "billiard_ball_projectile_small", fps: 12, scale: 1.5f, anchor: Anchor.MiddleCenter,
-            collidesWithProjectiles: true, collidesOnlyWithPlayerProjectiles: true))
+            collidesWithProjectiles: true, collidesOnlyWithPlayerProjectiles: true, customClip: true))
           .Attach<BilliardBall>()
           .Assign(out _BilliardBall);