+++ src/Cwaff-Guns/KALI.cs
@@ -24,7 +24,7 @@ public class KALI : CwaffGun
           .SetAttributes(quality: ItemQuality.B, gunClass: GunClass.CHARGE, reloadTime: 0.1f, ammo: 200, fireAudio: "kali_shoot_sound");
 
         _KaliProjectile = gun.InitProjectile(GunData.New(range: 9999f, force: 3f, cooldown: 0.1f, collidesWithProjectiles: true,
-            clipSize: 1, sprite: "kali_projectile", shootStyle: ShootStyle.Charged))
+            clipSize: 1, sprite: "kali_projectile", shootStyle: ShootStyle.Charged, customClip: true))
           .Attach<PierceProjModifier>(pierce => { pierce.penetration = 999; pierce.penetratesBreakables = true; });
         _KaliProjectile.AddTrailToProjectilePrefab("kali_trail", fps: 60, cascadeTimer: C.FRAME, softMaxLength: 1f, destroyOnEmpty: false,
           dispersalPrefab: Items.FlashRay.AsGun().DefaultModule.projectiles[0].GetComponentInChildren<TrailController>().DispersalParticleSystemPrefab);