+++ src/Cwaff-Guns/RCLauncher.cs
@@ -18,7 +18,7 @@ public class RCLauncher : CwaffGun
             loopReloadAt: 0, fireAudio: "rc_car_launch_sound", reloadAudio: "rc_car_reload_sound")
           .InitSpecialProjectile<RCGuidedProjectile>(GunData.New(sprite: "rc_car_projectile", clipSize: 7, cooldown: 0.1f,
             shootStyle: ShootStyle.SemiAutomatic, speed: 20f, damage: 9f, range: 9999f, pierceBreakables: true,
-            shouldRotate: false, shouldFlipHorizontally: false, shouldFlipVertically: false,
+            shouldRotate: false, shouldFlipHorizontally: false, shouldFlipVertically: false, customClip: true,
             spawnSound: "rc_car_engine_sound", stopSoundOnDeath: true, destroySound: "rc_car_crash_sound"))
           .Attach<RCGuidedProjectile>(igp => {
             igp.dumbfireTime          = 0.2f;