+++ src/Cwaff-Guns/Starmageddon.cs
@@ -19,7 +19,7 @@ public class Starmageddon : CwaffGun
             reloadFps: 30, loopReloadAt: 0)
           .InitProjectile(GunData.New(clipSize: 30, cooldown: 0.125f, angleVariance: 15.0f,
             shootStyle: ShootStyle.Automatic, damage: 8.0f, speed: 60.0f, range: 999999f, spawnSound: "starmageddon_fire_sound",
-            sprite: "starmageddon_bullet", fps: 12, scale: 0.5f, anchor: Anchor.MiddleCenter, useDummyChargeModule: true,
+            sprite: "starmageddon_bullet", fps: 12, scale: 0.5f, anchor: Anchor.MiddleCenter, useDummyChargeModule: true, customClip: true,
             // overrideColliderPixelSizes: new IntVector2(128, 128), //WARNING: large hitboxes apparently lag the game????
             shrapnelVFX: VFX.Create("starmageddon_shrapnel"), shrapnelCount: 5))
           .Attach<StarmageddonProjectile>();